Build System Update
The following provides an update to the build system to support
a target and host prefix. This patch also addressed issues with
the toolchain and the dependence on macros to make the code
compiler properly. Finally, this also updated the internal
variables to all include BAREFLANK_, which greatly improves the
ability to see which variables are ours vs. which variables
belong to CMake.